Refactor moderation event to only delete messages in a specific channel

+++ src/discord/events/attendance/moderation.ts
@@ -4,7 +4,7 @@ new Event({
     name: "Moderation",
     event: "messageCreate",
     run(message) {
-      console.log(message.content);
+      if(message.channel.id !== "1241637598271901778") return;
       if(!message.author.bot) {
         message.delete();
         return;
